==> 02_clean_and_merge/03_clean_newarticles_counts_frequencies.Rmd <==
---
title: "Preprocessing: counting countries Xinhua 2000-2014"
author: "Lucie Lu"
date: "10/6/2021"
output: html_document
---

```{r}

list.files("Xinhua_news_articles_2000-2014/")

Xinhua_2014 <- read.csv("Xinhua_news_articles_2000-2014/Xinhua_2014.csv")
#head(Xinhua_2000 )
library(tidyverse)
#https://hbctraining.github.io/Intro-to-R/lessons/08_intro_tidyverse.html
Xinhua_2014 <- Xinhua_2014 %>% 
  rownames_to_column() %>% 
  as_tibble()

#replace empty country with extracted locations
Xinhua_2014$country <- ifelse(Xinhua_2014$country==" ", Xinhua_2014$extracted_locations, Xinhua_2014$country)

```

```{r}
df <- Xinhua_2014  %>% 
  pull(country) %>% 
  str_split("[|]")

#max(lengths(df))

names(df) <- Xinhua_2014$rowname
#df_2 <- data.frame(matrix(unlist(df), nrow=max(lengths(df)), byrow=TRUE),stringsAsFactors=FALSE)

library (plyr)
df_2 <- ldply (df, data.frame)
df_2 <- tibble(df_2)
colnames(df_2)[2] <- "country_mentioned"

df_2$countries_mentioned <- as.factor(df_2$country_mentioned)
```

```{r}
library(dplyr)
new_df_2014 <- df_2 %>% 
  group_by(country_mentioned) %>%
  dplyr::summarise(n = n()) %>%
#  df_2[!(df_2$country_mentioned == "")]  %>%
#  mutate(freq = round(n/sum(n), 3)) %>%
  arrange(desc(n))

new_df_2014 <- new_df_2014 %>% mutate(year=2014)

#duplicated name summarise in both packages
#https://statisticsglobe.com/r-error-must-only-be-used-inside-dplyr-verbs
```

```{r}
# list all of the files here
new_df_2000
new_df_2001
new_df_2002
new_df_2003
new_df_2004
new_df_2005
new_df_2006
new_df_2007
new_df_2008 
new_df_2009
new_df_2010
new_df_2011
new_df_2012
new_df_2013
new_df_2014
```

```{r}
all_df <- rbind(new_df_2000,
new_df_2001,
new_df_2002,
new_df_2003,
new_df_2004,
new_df_2005,
new_df_2006,
new_df_2007,
new_df_2008,
new_df_2009,
new_df_2010,
new_df_2011,
new_df_2012,
new_df_2013,
new_df_2014)
str(all_df)
```

```{r}
#transform to wide data for better graphing
#https://cran.r-project.org/web/packages/data.table/vignettes/datatable-reshape.html
library(data.table)
data_all_df_long <- dcast(all_df, country_mentioned ~ year, value.var = "n")
head(data_all_df_long, 20)

```

```{r export data}

#write.csv(data_all_df_long, "Xinhua_2001_2014_countrycounts.csv")

```

```{r get frequency tables}

### try this one
all_df %>% 
  group_by(country_mentioned, year) %>%
#  dplyr::summarise(n2 = n()) %>%
  mutate(freq_perc = round(100 * n / sum(all_df[all_df$year==2000,]$n), digits=2))

#sum the values instead of counting
sum(all_df[all_df$year==2000,]$n)

###end of trying
#################
freq_perc_by_year <- matrix(NA, nrow=length(existing_years), ncol=1)

existing_years <- seq(2000, 2014, by=1)

for (i in 1:length(existing_years)){
  freq_perc_by_year[i] = sum(all_df[all_df$year==existing_years[i],]$n)
}

freq_perc_by_year <- cbind(freq_perc_by_year, existing_years)
colnames(freq_perc_by_year)[1] <- "total_n_by_year"
colnames(freq_perc_by_year)[2] <- "year"

```

```{r merge the two}

all_df2 <- dplyr::inner_join(all_df, data.frame(freq_perc_by_year), by='year')

all_df2 <- all_df2 %>% 
  group_by(country_mentioned, year) %>%
#  dplyr::summarise(n2 = n()) %>%
  mutate(freq_perc = round(100 * n / total_n_by_year, digits=2))

```

```{r frequency tables}
#transform to wide data for better graphing
#https://cran.r-project.org/web/packages/data.table/vignettes/datatable-reshape.html
library(data.table)
data_all_df2_long <- reshape2::dcast(all_df2, country_mentioned ~ year, value.var = c("freq_perc"))
head(data_all_df2_long, 20)

```

```{r export data frequency}

#write.csv(data_all_df2_long, "Xinhua_2001_2014_countryfrequencies.csv")

```
